Dev branch sync

+++ core/vtk/ttkTrackingFromFields/ttkTrackingFromFields.cpp
@@ -2,7 +2,7 @@
 
 vtkStandardNewMacro(ttkTrackingFromFields)
 
-  constexpr unsigned int str2int(const char *str, int h = 0) {
+  constexpr unsigned long long str2int(const char *str, int h = 0) {
   return !str[h] ? 5381 : (str2int(str, h + 1) * 33) ^ str[h];
 }
